Detect recent installations and exit 0
Since nodenv-doctor is run immediately after nodenv-installer, for many
installations (non homebrew), the doctor will _always_ fail because it's
still running in a shell that requires PATH to be updated.

Since many users want to use nodenv-installer in an automated/unattended
fashion, the installer/doctor should exit with a zero status code after
a "successful" installation.

However, the doctor won't consider a missing nodenv PATH to be
"successful", so quandry!

Now, the doctor will attempt to guess that an installation was "recent"
but determining if the nodenv bin was recently modified (< 1min). If so,
then we print an extra little message and exit 0 instead of 1.